+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/ui/SeasonsFragment.java
@@ -51,7 +51,7 @@ import com.battlelancer.seriesguide.provider.SeriesGuideContract.ListItemTypes;
 import com.battlelancer.seriesguide.provider.SeriesGuideContract.Seasons;
 import com.battlelancer.seriesguide.settings.DisplaySettings;
 import com.battlelancer.seriesguide.ui.dialogs.ManageListsDialogFragment;
-import com.battlelancer.seriesguide.ui.dialogs.SortDialogFragment;
+import com.battlelancer.seriesguide.ui.dialogs.SingleChoiceDialogFragment;
 import com.battlelancer.seriesguide.util.DBUtils;
 import com.battlelancer.seriesguide.util.EpisodeTools;
 import com.battlelancer.seriesguide.util.EpisodeTools.SeasonWatchedType;
@@ -517,7 +517,8 @@ public class SeasonsFragment extends ListFragment implements
 
     private void showSortDialog() {
         FragmentManager fm = getFragmentManager();
-        SortDialogFragment sortDialog = SortDialogFragment.newInstance(R.array.sesorting,
+        SingleChoiceDialogFragment sortDialog = SingleChoiceDialogFragment.newInstance(
+                R.array.sesorting,
                 R.array.sesortingData, mSorting.index(),
                 DisplaySettings.KEY_SEASON_SORT_ORDER, R.string.pref_seasonsorting);
         sortDialog.show(fm, "fragment_sort");