+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/ui/ShowsFragment.java
@@ -27,9 +27,11 @@ import android.preference.PreferenceManager;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.ActivityOptionsCompat;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
 import android.support.v4.app.LoaderManager;
 import android.support.v4.content.CursorLoader;
 import android.support.v4.content.Loader;
+import android.support.v4.view.ViewCompat;
 import android.text.format.DateUtils;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -49,11 +51,11 @@ import com.battlelancer.seriesguide.provider.SeriesGuideContract.Shows;
 import com.battlelancer.seriesguide.settings.AdvancedSettings;
 import com.battlelancer.seriesguide.settings.DisplaySettings;
 import com.battlelancer.seriesguide.settings.ShowsDistillationSettings;
+import com.battlelancer.seriesguide.ui.dialogs.SingleChoiceDialogFragment;
 import com.battlelancer.seriesguide.util.DBUtils;
-import com.battlelancer.seriesguide.util.FabAbsListViewScrollDetector;
 import com.battlelancer.seriesguide.util.ShowMenuItemClickListener;
 import com.battlelancer.seriesguide.util.Utils;
-import com.melnykov.fab.FloatingActionButton;
+import com.uwetrottmann.androidutils.AndroidUtils;
 
 /**
  * Displays the list of shows in a users local library with sorting and filtering abilities. The
@@ -136,19 +138,10 @@ public class ShowsFragment extends Fragment implements
 
         // setup grid view
         mGrid = (GridView) getView().findViewById(android.R.id.list);
+        // enable app bar scrolling out of view only on L or higher
+        ViewCompat.setNestedScrollingEnabled(mGrid, AndroidUtils.isLollipopOrHigher());
         mGrid.setAdapter(mAdapter);
         mGrid.setOnItemClickListener(this);
-
-        // setup floating action button for adding shows
-        final FloatingActionButton buttonAddShow = (FloatingActionButton) getView().findViewById(
-                R.id.buttonShowsAdd);
-        mGrid.setOnScrollListener(new FabAbsListViewScrollDetector(buttonAddShow));
-        buttonAddShow.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(getActivity(), AddActivity.class));
-            }
-        });
 
         // listen for some settings changes
         PreferenceManager
@@ -308,6 +301,28 @@ public class ShowsFragment extends Fragment implements
             getActivity().supportInvalidateOptionsMenu();
 
             fireTrackerEventAction("Filter Removed");
+            return true;
+        } else if (itemId == R.id.menu_action_shows_filter_upcoming_range) {
+            // yes, converting back to a string for comparison
+            String upcomingLimit = String.valueOf(
+                    AdvancedSettings.getUpcomingLimitInDays(getActivity()));
+            String[] filterRanges = getResources().getStringArray(R.array.upcominglimitData);
+            int selectedIndex = 0;
+            for (int i = 0, filterRangesLength = filterRanges.length; i < filterRangesLength; i++) {
+                String range = filterRanges[i];
+                if (upcomingLimit.equals(range)) {
+                    selectedIndex = i;
+                    break;
+                }
+            }
+
+            SingleChoiceDialogFragment upcomingRangeDialog = SingleChoiceDialogFragment.newInstance(
+                    R.array.upcominglimit,
+                    R.array.upcominglimitData,
+                    selectedIndex,
+                    AdvancedSettings.KEY_UPCOMING_LIMIT,
+                    R.string.pref_upcominglimit);
+            upcomingRangeDialog.show(getFragmentManager(), "upcomingRangeDialog");
             return true;
         } else if (itemId == R.id.menu_action_shows_sort_title) {
             if (mSortOrderId == ShowsDistillationSettings.ShowsSortOrder.TITLE_ID) {